requirements.txt
discord.py
python-dotenv
google-genai
Pillow
PyNaCl
pytesseract
